Extract serverId from garbage-collection
Code which runs very early in startup needs to be very careful about
minimizing imports and avoiding unintentional side-effects through
imported symbols living in the same file as side-effectful functions.

`configureLogger.ts` was importing serverId from garbage-collection very
early on in process startup.  This had the result of registering the
garbage-collection Meteor.startup() function very early in the process's
lifecycle, which has the effect of running the Servers insertion before
doing things like schema updates, index creation, or running migrations.

This is potentially problematic because it results in us issuing
mongo-schema-checked writes against a version of the schema that existed
in the database rather than the one that we would attach based on the
current version of the code.

This is especially problematic for the Servers collection because it
results in the server running, but no other servers in the cluster
knowing that it exists.  This prevents them from connecting over the
webrtc heartbeat channel, which leads the primary server instance to
believe that all the children it has spawned are unresponsive, so it
marks them for deletion 10 seconds after spawning them and continues to
do this in a loop indefinitely.
