+++ imports/server/serverId.ts
@@ -1,0 +1,5 @@
+import { Random } from "meteor/random";
+
+const serverId = Random.id();
+
+export default serverId;
